place documentation on each function
