+++ qiskit/tools/__init__.py
@@ -23,18 +23,7 @@ Parallel Routines
 .. autosummary::
    :toctree: ../stubs/
 
-parallel_map
-===============
-.. code-block::
-
-   import time
-   from qiskit.tools.parallel import parallel_map
-
-   def func(_):
-      time.sleep(0.1)
-      return 0
-
-   parallel_map(func, list(range(10)));
+   parallel_map
 
 Monitoring
 ==========
