+++ qiskit/tools/monitor/__init__.py
@@ -10,30 +10,7 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""A module for monitoring jobs, backends, etc.
-
-
-job_monitor
-===============
-
-.. code-block::
-
-    from qiskit import BasicAer, execute
-    from qiskit.circuit import QuantumRegister, ClassicalRegister, QuantumCircuit
-    from qiskit.tools.monitor import job_monitor
-    sim_backend = BasicAer.get_backend("qasm_simulator")
-
-    q = QuantumRegister(2)
-    c = ClassicalRegister(2)
-    qc = QuantumCircuit(q, c)
-
-    qc.h(q[0])
-    qc.cx(q[0], q[1])
-    qc.measure(q, c)
-    job_sim = execute(qc, backend=sim_backend)
-    job_monitor(job_sim)
-
-"""
+"""A module for monitoring jobs, backends, etc."""
 
 from .job_monitor import job_monitor
 from .overview import backend_monitor, backend_overview
