+++ qiskit/tools/parallel.py
@@ -137,6 +137,14 @@ def parallel_map(  # pylint: disable=dangerous-default-value
         terra.parallel.start: The collection of parallel tasks are about to start.
         terra.parallel.update: One of the parallel task has finished.
         terra.parallel.finish: All the parallel tasks have finished.
+
+    Example:
+           import time
+           from qiskit.tools.parallel import parallel_map
+           def func(_):
+                time.sleep(0.1)
+                return 0
+           parallel_map(func, list(range(10)));
     """
     if len(values) == 0:
         return []
